extracted function from file and moved to page.modules.js
